Checkpoint
Only Grabs necessary pages, but still does not allow user to navigate to next pages.
Idea on how to fix:
The PageList object needs to be as long as actual number of entries but with empty objects. Once user changes page size, navigates to new page, or changes something, new codes will be lazy loaded